ファイルに切り出す
