+++ lib/procon_bypass_man/procon/performance_measurement.rb
@@ -1,48 +1,11 @@
 module ProconBypassMan::Procon::PerformanceMeasurement; end
 
 require 'benchmark'
+require 'procon_bypass_man/procon/performance_measurement/measurements_summarizer'
 require 'procon_bypass_man/procon/performance_measurement/span_queue'
 require 'procon_bypass_man/procon/performance_measurement/queue_over_process'
 
 module ProconBypassMan::Procon::PerformanceMeasurement
-  class PerformanceMetrics < Struct.new(:time_taken_p50,
-                                        :time_taken_p95,
-                                        :time_taken_p99,
-                                        :time_taken_max,
-                                        :read_error_count,
-                                        :write_error_count); end
-
-  class MeasurementsSummarizer
-    def initialize(measurements: )
-      @measurements = measurements
-    end
-
-    # @return [PerformanceMetrics]
-    def summarize
-      sorted_time_taken = @measurements.map(&:time_taken).sort
-      time_taken_p50 = percentile(sorted_list: sorted_time_taken, percentile: 0.50)
-      time_taken_p95 = percentile(sorted_list: sorted_time_taken, percentile: 0.95)
-      time_taken_p99 = percentile(sorted_list: sorted_time_taken, percentile: 0.99)
-      time_taken_max = sorted_time_taken.last || 0
-      total_read_error_count = @measurements.map(&:read_error_count).sum
-      total_write_error_count = @measurements.map(&:write_error_count).sum
-      PerformanceMetrics.new(time_taken_p50, time_taken_p95, time_taken_p99, time_taken_max, total_read_error_count, total_write_error_count)
-    end
-
-    private
-
-    # @param [Array<any>]
-    # @param [Float] percentile
-    # @return [Float]
-    def percentile(sorted_list: , percentile: )
-      return 0.0 if sorted_list.empty?
-      values_sorted = sorted_list
-      k = ((percentile*(values_sorted.length-1))+1).floor - 1
-      f = ((percentile*(values_sorted.length-1))+1).modulo(1)
-      return(values_sorted[k] + (f * (values_sorted[k+1] - values_sorted[k]))).floor(3)
-    end
-  end
-
   class PerformanceSpan
     attr_accessor :time_taken
     attr_reader :write_error_count, :read_error_count
@@ -82,9 +45,9 @@ module ProconBypassMan::Procon::PerformanceMeasurement
   end
 
   # @param [MeasurementCollection] measurements
-  # @return [PerformanceMetrics]
+  # @return [ProconBypassMan::Procon::PerformanceMeasurement::MeasurementsSummarizer::PerformanceMetric]
   # jobから呼ばれる予定
   def self.summarize(measurements: )
-    MeasurementsSummarizer.new(measurements: measurements).summarize
+    ProconBypassMan::Procon::PerformanceMeasurement::MeasurementsSummarizer.new(measurements: measurements).summarize
   end
 end
