+++ lib/procon_bypass_man/procon/performance_measurement/span_queue.rb
@@ -16,6 +16,7 @@ class ProconBypassMan::Procon::PerformanceMeasurement::SpanQueue
   end
 
   # @param [PerformanceSpan] span
+  # bypassプロセスから呼ばれる
   def push(span)
     current_key = generate_bucket_key
